lib/visualize/pyproject.toml
[project]
name = "terrascore-visualize"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["matplotlib>=3.9", "numpy>=1.26", "Pillow>=10.0"]
